pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ghost-office-hunter"
version = "1.0.0"
description = "AI-powered compliance investigation tool for detecting ghost offices and shell companies"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Ghost Office Hunter Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Office/Business :: Financial",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "crewai==1.7.2",
    "crewai-tools==1.7.2",
    "ddgs==9.10.0",
    "python-dotenv==1.1.1",
]

[project.scripts]
ghost-hunter = "main:main"

[tool.setuptools]
packages = ["."]
